Partial support for https://github.com/jellyfin/jellyfin-roku/issues/36
Partial implemented support for the languages and dialets that Roku supports.

A number of structural issues encountered the biggest are:
1. Programatically catenating strings which assumes a particular language syntax.
2. Using display strings in buttons, etc. to determine action.
3. Building display objects in BrightScript rather than in XML. This results in many cases where you have to translate the same string many times (e.g. "Cancel")

Side note: My editor is set to remove trailing spaces and to assure a new line at the end of the file. Existing source code seems to vary on this. I can revert those few changes if that is a problem.

Finally, added https://reuse.software/spec/ compliant license information to the new files and the files I touched.

Signed-off-by: Tod Fitch <tod@fitchfamily.org>